# routes/products.js
// 跨境发货管理系统 - 产品路由

const express = require('express');
const { authenticateToken, authorizeRoles, authorizeResourceAccess } = require('../utils/auth');
const { body } = require('express-validator');

const router = express.Router();

// 获取产品列表
router.get('/', authenticateToken, async (req, res) => {
  try {
    const { limit = 10, offset = 0, search } = req.query;
    
    // 这里应该连接到实际的数据库查询
    // 示例响应
    res.json({
      products: [],
      total: 0
    });
  } catch (error) {
    console.error('获取产品列表失败:', error);
    res.status(500).json({ error: '服务器内部错误' });
  }
});

// 获取特定产品
router.get('/:id', authenticateToken, authorizeResourceAccess('product'), async (req, res) => {
  try {
    const { id } = req.params;
    
    // 这里应该连接到实际的数据库查询
    // 示例响应
    res.json({ id, name: '示例产品' });
  } catch (error) {
    console.error('获取产品失败:', error);
    res.status(500).json({ error: '服务器内部错误' });
  }
});

// 创建新产品
router.post('/', authenticateToken, authorizeRoles('admin', 'sales'), [
  body('name').notEmpty(),
  body('price').isNumeric(),
  body('cost').isNumeric()
], async (req, res) => {
  try {
    const productData = req.body;
    productData.created_by = req.user.id;
    
    // 这里应该连接到实际的数据库插入
    // 示例响应
    res.status(201).json(productData);
  } catch (error) {
    console.error('创建产品失败:', error);
    res.status(500).json({ error: '服务器内部错误' });
  }
});

// 更新产品
router.put('/:id', authenticateToken, authorizeRoles('admin', 'sales'), authorizeResourceAccess('product'), [
  body('name').optional().notEmpty(),
  body('price').optional().isNumeric(),
  body('cost').optional().isNumeric()
], async (req, res) => {
  try {
    const { id } = req.params;
    const productData = req.body;
    
    // 这里应该连接到实际的数据库更新
    // 示例响应
    res.json({ id, ...productData });
  } catch (error) {
    console.error('更新产品失败:', error);
    res.status(500).json({ error: '服务器内部错误' });
  }
});

// 删除产品
router.delete('/:id', authenticateToken, authorizeRoles('admin'), authorizeResourceAccess('product'), async (req, res) => {
  try {
    const { id } = req.params;
    
    // 这里应该连接到实际的数据库删除
    // 示例响应
    res.status(200).json({ message: '删除成功' });
  } catch (error) {
    console.error('删除产品失败:', error);
    res.status(500).json({ error: '服务器内部错误' });
  }
});

module.exports = router;